+++ src/main/java/de/codescape/jira/plugins/scrumpoker/upgrade/Upgrade09CreateDefaultCardSet.java
@@ -3,10 +3,12 @@ package de.codescape.jira.plugins.scrumpoker.upgrade;
 import com.atlassian.plugin.spring.scanner.annotation.export.ExportAsService;
 import com.atlassian.sal.api.upgrade.PluginUpgradeTask;
 import de.codescape.jira.plugins.scrumpoker.model.GlobalSettings;
-import de.codescape.jira.plugins.scrumpoker.model.SpecialCards;
 import de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSettingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+
+import static de.codescape.jira.plugins.scrumpoker.model.SpecialCards.COFFEE_CARD;
+import static de.codescape.jira.plugins.scrumpoker.model.SpecialCards.QUESTION_MARK;
 
 /**
  * Create the simplified Fibonacci card set as the default card set.
@@ -42,7 +44,7 @@ public class Upgrade09CreateDefaultCardSet extends AbstractUpgradeTask {
     }
 
     private String simplifiedFibonacciCardSet() {
-        return SpecialCards.QUESTION_MARK + ", " + SpecialCards.COFFEE_CARD + ", 0, 1, 2, 3, 5, 8, 13, 20, 40, 100";
+        return QUESTION_MARK + ", " + COFFEE_CARD + ", 0, 1, 2, 3, 5, 8, 13, 20, 40, 100";
     }
 
 }