+++ docs/changelog.md
@@ -28,6 +28,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: require mandatory fields in global configuration
 * fix: ignore required fields when switching back to default settings
 * refactor: simplify way to identify pressed button for chosen action
+* refactor: static import card constants for readability
 
 ### [4.9.0] - 2020-02-16